requirements.txt
# FastAPI 관련 패키지
fastapi
uvicorn

# SQLAlchemy (MySQL 연결)
sqlalchemy

# 데이터베이스 관련 (MySQL + MongoDB)
pymongo
mysql-connector-python
pymysql

# 환경 변수 관리
python-dotenv

# 데이터 처리
pandas
numpy

# UUID 생성
uuid

# Scikit-learn (TF-IDF 벡터라이저)
scikit-learn

# 텍스트 전처리 (한국어 형태소 분석기)
kiwipiepy

# 로깅 및 기타 유틸
datetime

# 명령어 인자 파싱
argparse

pydantic[email]

starlette
alembic
mysqlclient
mecab-ko-dic